+++ src/dymoprint/__init__.py
@@ -1,4 +1,4 @@
-from .labeler import DymoLabeler
-from .metadata import __version__
+from dymoprint.lib.labeler import DymoLabeler
+from dymoprint.metadata import __version__
 
 __all__ = ["__version__", "DymoLabeler"]
