+++ src/dymoprint/gui/gui.py
@@ -21,9 +21,9 @@ from PyQt6.QtWidgets import (
 )
 from usb.core import NoBackendError, USBError
 
-from dymoprint.constants import DEFAULT_MARGIN_PX, ICON_DIR
-from dymoprint.detect import detect_device
-from dymoprint.dymo_print_engines import DymoRenderEngine, print_label
+from dymoprint.lib.constants import DEFAULT_MARGIN_PX, ICON_DIR
+from dymoprint.lib.detect import detect_device
+from dymoprint.lib.dymo_print_engines import DymoRenderEngine, print_label
 
 from .q_dymo_labels_list import QDymoLabelList
