+++ src/dymoprint/lib/barcode_writer.py
